requirements.txt.txt
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.3.3  # or your preferred database driver
bcrypt==4.0.1
pyjwt==2.7.0